fix parseFloat

+++ main.js
@@ -33,7 +33,7 @@ function myalertDiv(){
 function myalertBMI(){
     var num3 = document.getElementById("num3").value
     var num4 = document.getElementById("num4").value
-    var sum = parseInt(num3) / ((parseInt(num4)/100)*(parseInt(num4)/100))
+    var sum = parseFloat(num3) / ((parseInt(num4)/100)*(parseInt(num4)/100))
     document.getElementById("result").innerHTML = "แสดงผลBMI : " + sum.toFixed(2)
     alert(sum)
 }
